+++ console.js
@@ -4,9 +4,6 @@ const path = require('path');
 const repl = require('repl');
 const figlet = require('figlet');
 const DLogsAPI = require('./DLogsAPI.js');
-const {app, BrowserWindow, ipcMain} = require('electron');
-const url = require('url');
-
 const dlogs = new DLogsAPI(3000, '127.0.0.1',
     {
        "appName": "DLogs",
@@ -77,58 +74,36 @@ const ASCII_Art = (word) => {
         return new Promise(_aa);
 }
 
-// electron window global object
-let win;
+// Handling promises in REPL (for node < 10.x)
+const replEvalPromise = (cmd,ctx,filename,cb) => {
+  let result=eval(cmd);
+  if (result instanceof Promise) {
+    return result.then(response=>cb(null,response));
+  }
+  return cb(null, result);
+}
 
-const createWindow = () => {
-	  // Create the browser window.
-	  __master_init(masterpass).then((dlogs) => {
-	    win = new BrowserWindow({minWidth: 1280, minHeight: 960, resizable: true, icon: path.join(__dirname, 'public', 'assets', 'icon', '11be_logo.png')});
-	    win.setMenu(null);
-	
-	    // and load the index.html of the app.
-	    win.loadURL(url.format({
-	      pathname: path.join(__dirname, '/public/index.html'),
-	      protocol: 'file:',
-	      slashes: true
-	    }))
-	
-	    global.dlogs = dlogs;
-	
-	    // Open the DevTools.
-	    win.webContents.openDevTools()
-	
-	    // Emitted when the window is closed.
-	    win.on('closed', () => {
-	      // Dereference the window object, usually you would store windows
-	      // in an array if your app supports multi windows, this is the time
-	      // when you should delete the corresponding element.
-	      win = null
-	    })
-	  })
-	}
-
-	app.on('ready', createWindow)
-
-	// Whole process reloader via ipcRenderer for config reload
-	ipcMain.on('reload', (e, args) => {
-	        app.relaunch();
-	        app.exit();
-	});
-	
-	// Quit when all windows are closed.
-	app.on('window-all-closed', () => {
-	  // On macOS it is common for applications and their menu bar
-	  // to stay active until the user quits explicitly with Cmd + Q
-	  if (process.platform !== 'darwin') {
-	    app.quit()
-	  }
+// REPL main function
+const terminal = (slogan = 'ElevenBuckets :  BladeIron') => {
+  return __master_init(masterpass).then((dlogs) => {
+		 return ASCII_Art(slogan)
 	})
-	
-	app.on('activate', () => {
-	  // On macOS it's common to re-create a window in the app when the
-	  // dock icon is clicked and there are no other windows open.
-	  if (win === null) {
-	    createWindow()
-	  }
-	})
+        .then((art) => {
+          console.log(art + "\n");
+
+          let r = repl.start({ prompt: '[-= ElevenBuckets@Web3_Summit_2018 =-]$ ', eval: replEvalPromise });
+          r.context = {dlogs};
+
+          r.on('exit', () => {
+                  console.log("\n" + 'Stopping CLI...');
+                  process.exit(0);
+          })
+    })
+    .catch((err) => {
+        console.log(err);
+        process.exit(12);
+    })
+}
+
+// Main
+terminal('DLogs  by  ElevenBuckets');
